Add provider configuration support
This allows users to specify a supported provider for their modem. If
they choose to use a custom modem that does not use a supported provider
then they can add that to the `:extra_modems` config with out a provider
given.

This is useful for really custom chatscripts and modem configurations.
However, this also starts to provide a supported group of
modem-provider pairings that is easily configured.