Quectel: only connect to the first provider
This was, in fact, what was happening, but it didn't look like it. While
multiple PDP contexts were being configured, the PPP connection line was
hardcoded to connect to use the first one. This update still allows
multiple providers to be specified (for backwards/forwards
compatibility), but only configures the first one.