Allow modem implementations to normalize configs
This is necessary for them to have custom options since
`VintageNetMobile` wouldn't know how to normalize those. A side effect
of this change is that configuration errors will be caught earlier at
normalize-time rather than when converted to raw configurations. This
may have benefits in the future since normalization tends to be called
early in the configuration process.