Refactor AT parser to be more general
This adds a small Leex-generated lexer to support parsing more kinds of
AT notifications. Errors also get handled now by returning error tuples.
This avoids crashing the monitoring GenServers when a glitch happens
when running the commands or there's a misformed notification that
somehow gets reported by the modem. Previously when this happened, a
GenServer would crash and the ppp session would be bounced. This seemed
harsh.