Fix pppd shim to work with newer glibc versions
This updates the shim to override a the newer symbol used for stat when
running newer versions of glibc.

This was found when watching the pppd scripts successfully connect, but
the notifications didn't get to Elixir. Erlinit's test fixture is a much
more complicated shim than the one here and it had the fix to this.